modem_fw: call functions to enable/disable HSI port
BZ: 33956

This patch for Droidboot is an adaptation of patches 43256 and 43269
made for Recovery OS.
HSI driver itself initializes the modem for Clovertrail (reset...).
Function calls have been added to manage this initialization. These
functions do nothing yet for Medfield.
Ports used to flash the modem depend on the board:
- /dev/ttyMFD1 (Uart) and /dev/ttyIFX0 (HSI) for Medfield/Lexington
- /dev/ttyIFX1 (HSI) for CLovertrail
Timeout have also been reduced to 10s (could be as low as 7s/1s
on Medfield).

Change-Id: I95961230635c702189f39c229d9815a3e741c4af
Signed-off-by: Christian Baritel <christianx.baritel@intel.com>
Reviewed-on: http://android.intel.com:8080/50926
Reviewed-by: Afantenos, Marie CecileX <marie.cecilex.afantenos@intel.com>
Reviewed-by: Roulliere, Pierre <pierre.roulliere@intel.com>
Reviewed-by: Alagnou, GregoryX <gregoryx.alagnou@intel.com>
Tested-by: Afantenos, Marie CecileX <marie.cecilex.afantenos@intel.com>
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>