libintelprov: New command fastboot flash radio_fuse_only
BZ: 53917

The command "fastboot flash radio_fuse_only" get the modem chip
fusing parameters and store it in the file
[/data]/logs/modem_fuse.fus.
No modem flashing is made with this command.
The firmware file as a parameter is still needed to find into it
the model of the modem (6260, 6360, 7160, ...).
The command "fastboot flash radio_fuse" still exists and flash
the modem after getting the fusing parameters.

Change-Id: I9837da15f4bed5dcf4735871c1beb4dc8c12e784
Signed-off-by: Christian Baritel <christianx.baritel@intel.com>
Reviewed-on: http://android.intel.com:8080/69796
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>