Control IFWI downgrade if PROD or DEV or DEBUG
BZ: 41070

The goal is to allow IFWI downgrade only if the ifwi is a DEV
or DEBUG one.
if the IFWI is an end user or a PROD one the downgrade is not
allowed.
The IFWI type is checked by parsing the ifwi binary and getting
PTI field value.
If PTI bit 7 located at ofsset 0x030C in SMIP area is 1 this
corresponds to a DEV or DEBUG IFWI. The downgrade is allowed.
If the PTI bit 7 is 0 this corresponds to a PROD IFWI. The
downgrade is then forbidden.

Change-Id: Ife6ede86a9e6d93596db19683ae9f80f6a54da8f
Signed-off-by: Michel Jauffres <michelx.jauffres@intel.com>
Reviewed-on: http://android.intel.com:8080/70697
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Betous, GuillaumeX <guillaumex.betous@intel.com>
Reviewed-by: Pince, Jean-christophe <jean-christophe.pince@intel.com>
Reviewed-by: Imberton, Guilhem <guilhem.imberton@intel.com>
Tested-by: Touzeau, GabrielX <gabrielx.touzeau@intel.com>
Reviewed-by: lab_cactus <lab_cactus@intel.com>
Tested-by: lab_cactus <lab_cactus@intel.com>