Droidboot: support for flashing IFWI using fastboot commands
BZ: 73560

In Tangier, since there is no support for FOTA, the IFWI cannot be flashed
via IPC commands to the SCU. This code change in the userspace adds the ability
to flash the IFWI by changing the boot partition to read-write, and writing to
it from IA directly while the device is in fastboot mode. Because of this, the
DnX is not required to be flashed for TNG. This ability requires a change in
the kernel, which is submitted in a separate patch.

Change-Id: Iaa653079e2e0bab974cf3ea66d527a55934c578f
Signed-off-by: Vivek Eswar <vivek.eswar@intel.com>
Reviewed-on: http://android.intel.com:8080/84746
Reviewed-by: Ponnusamy, Siva Prasath <siva.prasath.ponnusamy@intel.com>
Tested-by: Ponnusamy, Siva Prasath <siva.prasath.ponnusamy@intel.com>
Reviewed-by: Romieu, Benoit <benoit.romieu@intel.com>
Tested-by: Romieu, Benoit <benoit.romieu@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>