Allow FOTA update from r4.1 to r4 main
BZ: 75047

The goal of this patch is to allow updating from a installed r4.1
build to r4 main.
At boot time the updater script in the ota zip file is executed.
It asserts to ro.build.device or ro.build.product with its build
property.
As the build type has changed between the r4.1 branch (for example
mfld_pr2) and the r4 main (for example blackbay), the patch consists
in adding a test that verifies the ro.build.device or ro.build.product
with the old build name.
Verification is implemented in the libintelprov release tool python
scripts used to construct the ota zip and based on a file named
product_name_mapping.def that lists the old build types compatibles
with the new ones.

Notice that if the Fota application is used to download the package
it must be able to accept the ota package with the new naming
convention. This is handled by BZ75039.

Change-Id: Id931d77e9577d877e21ba89503b4f2d9ec5b92c7
Signed-off-by: Michel Jauffres <michelx.jauffres@intel.com>
Reviewed-on: http://android.intel.com:8080/85367
Reviewed-by: Pince, Jean-christophe <jean-christophe.pince@intel.com>
Tested-by: Afantenos, Marie Cecile <marie.cecile.afantenos@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>