Add alternate fw_update path for dnx/ifiwi update for kernel 3.4
BZ: 86441

Due to new kernel 3.4 the intel_fw_update.c creates a new sysfs
interface.
dnx, fw_info and ifwi are located in /sys/kernel/fw_update/.

Current libintelprov looks in /sys/devices/ipc/intel_fw_update.0/
when the dnx and ifwi need to be updated.

This patch consists in adding in libintelprov the path to
/sys/kernel/fw_update/ as alternate in case /sys/devices/ipc/...
path cannot be accessed.

Change-Id: Id996c6652278b977aeca64baa248bc3fde3c2c69
Signed-off-by: Michel Jauffres <michelx.jauffres@intel.com>
Reviewed-on: http://android.intel.com:8080/91724
Reviewed-by: Castelain, DavidX <davidx.castelain@intel.com>
Reviewed-by: Betous, GuillaumeX <guillaumex.betous@intel.com>
Reviewed-by: Pince, Jean-christophe <jean-christophe.pince@intel.com>
Tested-by: Chotard, Celine <celine.chotard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>