libintelprov: test the return of cmfwdl_destroy_instance()
BZ: 90320

Change the code to test the return of the function
cmfwdl_destroy_instance() to verify that the cold reset has been
made correctly and that the modem is in IPC Ready state.
As cmfwdl_destroy_instance() wait for the modem to be in the IPC
Ready state, the final sleep(8) is no longer necessary.

Change-Id: Id78cf4d5b8c98a7daa71369289ceed4b29f800a0
Signed-off-by: Christian Baritel <christianx.baritel@intel.com>
Reviewed-on: http://android.intel.com:8080/95176
Reviewed-by: Rodrigues, JoseX <josex.rodrigues@intel.com>
Reviewed-by: Marmounier, Edward <edward.marmounier@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>