Force result of flash token command for dummy token.
BZ: 102042

If a dummy token, actually too small token, is received, the
flash_token function assumes that this is for development.
This allows to use the default flash_fru.xml with the Phone
Flash Tool, without preventing to execute the 'fastboot oem
fru set' command when in a development environment.

Change-Id: Id8075ff65e027f948bde20ad211082d36d104b3c
Signed-off-by: Florent Auger <florent.auger@intel.com>
Reviewed-on: http://android.intel.com:8080/109778
Reviewed-by: Chouleur, Sylvain <sylvain.chouleur@intel.com>
Reviewed-by: Imberton, Guilhem <guilhem.imberton@intel.com>
Reviewed-by: Jauffres, MichelX <michelx.jauffres@intel.com>
Tested-by: Molero Rodenas, BeatrizX <beatrizx.molero.rodenas@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>