Stub boot command returning OKAY status
BZ: 115703

Blankphone flashfile of BYT assume fastboot dnx mode will be used to
boot droidboot. For that, "fastboot boot" command is used, and then
"fastboot continue", then blankphone commands can operate.

But we need to handle the case where sdcard will be used to run
droidboot instead of fastboot dnx mode, but using the same blankphone
flashfile.

Android's legacy "fastboot boot" command is stubbed but is returning a
failure status. This need to be changed into a success status to avoid
interrupting phoneflashtool.
So boot command is overwritten to be stubbed and returning success status.

Change-Id: I9dd1c707814b8a262b38a261104ab02a75591e72
Signed-off-by: Sylvain Chouleur <sylvain.chouleur@intel.com>
Reviewed-on: http://android.intel.com:8080/114165
Reviewed-by: Tinelli, Vincent <vincent.tinelli@intel.com>
Reviewed-by: Afantenos, Marie Cecile <marie.cecile.afantenos@intel.com>
Tested-by: Molero Rodenas, BeatrizX <beatrizx.molero.rodenas@intel.com>
Tested-by: Tinelli, Vincent <vincent.tinelli@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>