[REVERTME] libintelprov: dynamically detect OSIP or FULL-GPT method
BZ: 146480

During the transition period between FDK and UEFI, it would be a lot
more convenient, even necessary in some cases, if libintelprov is able
to flash either OSIP images or FULL-GPT mode images. The
implementation is based on a simple heuristic, which is strong enough
for now, but as we never know if it will actually last long this patch
is repertoried as REVERTME and should be reverted in a few monthes
when the transition will be over.

Change-Id: I0bc5ac758ed89b326a1319b1c0b0afe49f1fe26e
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: http://android.intel.com:8080/139642
Reviewed-by: Tinelli, Vincent <vincent.tinelli@intel.com>
Tested-by: Lumbierres, PhilippeX <philippex.lumbierres@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>