recovery: add partitioning support
BZ: 147535

This patch provides a new Edify function which flash a partition
scheme. This is a very risky operation as we cannot reload the
partition scheme during an OTA update (before Android KitKat) and thus
the block device nodes in /dev/block won't be accurately representive
of the actual situation anymore. It would work only if the
modifications does not imply a global partition shift where the data
or the system partition would be shifted for instance.

Change-Id: Icc48d5ab8db44ae1b6b8a583f8a9d4ac7d84d276
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>