libintelprov: flash BOM token prior to any thing else
BZ: 153596

This patch adds a new function flash_bom_token that can be called
from edify script. It only flashes BOM token, and it is called
at very begining of flashing. If it fails (BOM token rejected by
chaabi) flashing is stopped before OSIP is invalidated, so that next
reboot will be in MOS, and no component is updated.

Change-Id: Id0ee1b0691eedffcf998b6866b830bcb200f404a
Signed-off-by: Gabriel Touzeau <gabrielx.touzeau@intel.com>