fastboot oem config command (partition based)
BZ: 167715

Implementation of a new fastboot OEM command

It writes the content of the only paramter in a file in /mnt/config.
This parameter will be used later at boot time to customize init.rc flow.

It's intented to be used with after a command "fastboot oem mount" in order
to have config partition mounted in /mnt/config.

Change-Id: I5c1fa383ca8e57ddfebc96388e11fafc714ab024
Signed-off-by: Guillaume Betous <guillaumex.betous@intel.com>
Signed-off-by: Laurent FERT <laurent.fert@intel.com>