ota_from_targetfiles use device option
BZ: 172702

now ota_from_targetfiles script uses call option,
those options are set depending on variable environment.
Those variable are not available for fotagen that call the otafromtargetfile
to generate incremental fota.

Using specific device option will solve the issue
since those option will be contain in the targetfile
that is use to point on a build to generate fota.

Change-Id: I5ac9e9ca7faa66467bdc5e8d8ff83adef6901041
Signed-off-by: Marie-Cecile Afantenos <marie.cecile.afantenos@intel.com>