oem: fix race condition between start_partitioning and partition cmds
BZ: 179075

Currently 'start_partitioning' is used to set a property named
sys.partitioning, if 'partition' command is used right after, there
is a chance that the property might not be set yet and hence
blankphone flashing will fail.

This is solved by waiting for a fixed time until the property is set
or returning error otherwise.

Change-Id: I9fbbf658c45432f86256dfd3686297cc6b4cb541
Signed-off-by: Omar Ramirez Luna <omar.r.ramirez.luna@intel.com>
Signed-off-by: Benoit Fradin <benoit.fradin@intel.com>
Signed-off-by: yesheng xu <yesheng.xu@intel.com>