diskd: Initial commit
BZ: 184400

diskd is a daemon that manage the /dev/disk/ directory.  It creates and
maintains a tree of symlinks that point out to the appropriate block
nodes that the init process creates in the /dev/block/ directory.

At its init stage, diskd discovers the block devices by exploring the
/sys/block directory.  Afterward, it listens the uevents to maintains
its tree.

Change-Id: I0e3f3d65e19ff46a3fea362f8ef1f7f3bbfce90a
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>