diskd support while still maintaining the parlink compatibility
BZ: 185008

In order to smooth the transition to diskd, we have to maintain the
backward compatibility with the partlink mechanism which is massively
deployed.  diskd is not yet available on all the devices and we have
to ensure that the FOTA update won't failed.

This patch implements a fallback mechanism : it tries to construct a
diskd formatted path, if it fails it then tries a partlink formatted
path.

Change-Id: I545d45411a2068b145d19697a2a2d565b8b6e44d
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>